requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-filter==23.5
psycopg2-binary==2.9.9
pillow==10.0.0
gunicorn==21.2.0
dj-database-url==2.1.0
redis==5.0.1